MinGW64 warnings muted.  #1439

+++ src/genmkvpwd.c
@@ -221,7 +221,7 @@ int main(int argc, char * * argv)
 		for(max_len=6;max_len<20;max_len++)
 		{
 			nbparts = mem_alloc(256*(max_lvl+1)*sizeof(long long)*(max_len+1));
-			printf("len=%u (%lu KB for nbparts) ", max_len, 256UL*(max_lvl+1)*(max_len+1)*sizeof(long long)/1024);
+			printf("len=%u (%lu KB for nbparts) ", max_len, (unsigned long)(256UL*(max_lvl+1)*(max_len+1)*sizeof(long long)/1024));
 			memset(nbparts, 0, 256*(max_lvl+1)*(max_len+1)*sizeof(long long));
 			nb_parts(0, 0, 0, max_lvl, max_len);
 			if(nbparts[0] > 1000000000)
@@ -242,7 +242,7 @@ int main(int argc, char * * argv)
 		for(max_lvl=100;max_lvl<=MAX_MKV_LVL;max_lvl++)
 		{
 			nbparts = mem_alloc(256*(max_lvl+1)*sizeof(long long)*(max_len+1));
-			printf("lvl=%u (%lu KB for nbparts) ", max_lvl, 256UL*(max_lvl+1)*(max_len+1)*sizeof(long long)/1024);
+			printf("lvl=%u (%lu KB for nbparts) ", max_lvl, (unsigned long)(256UL*(max_lvl+1)*(max_len+1)*sizeof(long long)/1024));
 			memset(nbparts, 0, 256*(max_lvl+1)*(max_len+1)*sizeof(long long));
 			nb_parts(0, 0, 0, max_lvl, max_len);
 			if(nbparts[0] > 1000000000)
@@ -263,7 +263,7 @@ int main(int argc, char * * argv)
 	}
 
 	nbparts = mem_alloc(256*(max_lvl+1)*sizeof(long long)*(max_len+1));
-	fprintf(stderr, "allocated %lu KB for nbparts\n", 256UL*(max_lvl+1)*(max_len+1)*sizeof(long long)/1024);
+	fprintf(stderr, "allocated %lu KB for nbparts\n", (unsigned long)(256UL*(max_lvl+1)*(max_len+1)*sizeof(long long)/1024));
 	memset(nbparts, 0, 256*(max_lvl+1)*(max_len+1)*sizeof(long long));
 
 	nb_parts(0, 0, 0, max_lvl, max_len);
